% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefi-package.R
\docType{package}
\name{cefi}
\alias{cefi-package}
\alias{cefi}
\title{cefi package}
\description{
R Tools for accessing the NOAA/PSL Climate Ecosystems and Fisheries Initiative Portal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/cefi}
}

}
