% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_active}
\alias{cefi_active}
\title{Retrieve one or more names of the active grids}
\usage{
cefi_active(x)
}
\arguments{
\item{x}{the tidync object (possibly pre-filtered)}
}
\value{
character vector of active variable names
}
\description{
Retrieve one or more names of the active grids
}
