% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_aggregate}
\alias{cefi_aggregate}
\title{Aggregate a dimension in a tidync_data object}
\usage{
cefi_aggregate(x, name = "member", fun = mean, na.rm = TRUE, nan.rm = TRUE)
}
\arguments{
\item{x}{tidync_data object}

\item{name}{chr the name of the dimension to aggregate}

\item{fun}{the function used to aggregate (if NULL then the input is returned)}

\item{na.rm}{logical, if TRUE pass the na.rm argument along to \code{fun}}

\item{nan.rm}{if TRUE then purge the output of NaN}
}
\value{
a tidync_data class object with reduced dimensionality
}
\description{
Aggregate a dimension in a tidync_data object
}
