% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_check_size}
\alias{cefi_check_size}
\alias{cefi_size}
\title{Check the size of the requested data object and warn if exceeds a specified value}
\usage{
cefi_check_size(x, warning_size = 200, bytes_per_element = 64)

cefi_size(x, bytes_per_element = 64)
}
\arguments{
\item{x}{tidync or tidync_data object}

\item{warning_size}{num, the threshold in MB above which to trigger a warning}

\item{bytes_per_element}{num, the number of bytes per element}
}
\value{
the input object
}
\description{
Check the size of the requested data object and warn if exceeds a specified value

Retrieve the size of the object in MB
}
