% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_dim}
\alias{cefi_dim}
\title{Retrieve dimensions for \code{tidync} or \code{tidync_data} objects.}
\usage{
cefi_dim(x)
}
\arguments{
\item{x}{tidync or tidync_data (from hyper_array)}
}
\value{
a named list of dimension lengths
}
\description{
Note that for \code{tidync} objects it is the dimensions fo the selected data
}
