% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_is_ensemble}
\alias{cefi_is_ensemble}
\title{Test if a cefi object is an ensemble}
\usage{
cefi_is_ensemble(x = cefi_open())
}
\arguments{
\item{x}{tidync or tidync_data object}
}
\value{
logical TRUE if the object is an ensemble
}
\description{
Test if a cefi object is an ensemble
}
