% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_reset}
\alias{cefi_reset}
\title{A silly function to reset the filters}
\usage{
cefi_reset(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{other arguments for hyper_transform}
}
\value{
the input object with the filters reset
}
\description{
A silly function to reset the filters
}
