% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_subset}
\alias{cefi_subset}
\title{Subset a CEFI tidync_data object}
\usage{
cefi_subset(x, i)
}
\arguments{
\item{x}{a tidync_data object}

\item{i}{indices (or name(s)) of one or more elements to subset}
}
\value{
a tidync_data object of the selected element(s)
}
\description{
Subset a CEFI tidync_data object
}
