% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_time}
\alias{cefi_time}
\title{Get the transformed time dimension, add a POSIXct time variable}
\usage{
cefi_time(x = cefi_open(), form = c("POSIXct", "Date")[2])
}
\arguments{
\item{x}{tidync object OR a tibble of transformed time}

\item{form}{chr, one of "POSIXct" or "Date" which determines the class of the
result}
}
\value{
tibble of time transform
}
\description{
Get the transformed time dimension, add a POSIXct time variable
}
