% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_var}
\alias{cefi_var}
\title{Get a CEFI variable as either a 'tidync_data' or 'stars' object}
\usage{
cefi_var(
  x = cefi_open(),
  var = cefi_active(x),
  form = c("tidync_data", "stars")[2],
  check_size = TRUE,
  ...
)
}
\arguments{
\item{x}{the tidync object (possibly pre-filtered)}

\item{var}{chr, the variable to retrieve}

\item{form}{one of 'tidync_data' or 'stars'}

\item{check_size}{logical, if TRUE check the request size and warn as needed}

\item{...}{other arguments passed through}
}
\value{
either 'tidync_data' or 'stars' object
}
\description{
Get a CEFI variable as either a 'tidync_data' or 'stars' object
}
