% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_attrs}
\alias{get_attrs}
\alias{set_attrs}
\alias{append_attr}
\alias{get_attr}
\alias{set_attr}
\title{Get and set attributes}
\usage{
get_attrs(x, namespace = "cefi")

set_attrs(x, value, namespace = "cefi")

append_attr(x, name, value, namespace = "cefi")

get_attr(x, name, namespace = "cefi")

set_attr(x, name, value, namespace = "cefi")
}
\arguments{
\item{x}{object that may have attributes}

\item{namespace}{chr, this is the name of list that holds cefi attributes}

\item{value}{the value to set as the attribute.  When using \code{set_attrs} it should be
a named list.}

\item{name}{chr, this is the name of and element of the cefi attribute namespace}
}
\value{
\code{get_attrs} returns a list (or NULL), \code{get_attr} returns a value (or NULL)
\code{set_attrs}, \code{set_attr} and \code{append_attr} return the updated input invisibly
}
\description{
Get and set attributes
}
