% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{get_static}
\alias{get_static}
\title{Get the static attribute}
\usage{
get_static(x)
}
\arguments{
\item{x}{tidync object}
}
\value{
the input's static tidync object
}
\description{
Get the static attribute
}
