% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_index}
\alias{nc_index}
\title{Retrieve ncdf4 style indices for start and count}
\usage{
nc_index(x)
}
\arguments{
\item{x}{tidync object}
}
\value{
list of start and count
}
\description{
Retrieve ncdf4 style indices for start and count
}
