% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{open_cefi}
\alias{open_cefi}
\title{Open a CEFI connection}
\usage{
open_cefi(x = dplyr::slice(read_catalog(), 1))
}
\arguments{
\item{x}{a (single-row) table of "CEFI_catalog" class}
}
\value{
a ncdf4 object
}
\description{
Open a CEFI connection
}
