% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{parse_cefi_opendap}
\alias{parse_cefi_opendap}
\title{Parse a cefi opendap filename}
\usage{
parse_cefi_opendap(
  uri = c("btm_o2.nwa.full.hcast.daily.raw.r20230401.199301-201912.nc",
    "thetao.nep.full.ss_refcast.monthly.raw.r20230401.ens_stats.i199303.nc",
    "tos.nwa.full.ltm_proj.yearly.raw.r20230401.proj_ssp585.enss.202001-209912.nc")
)
}
\arguments{
\item{uri}{chr one or more opendap uris}
}
\value{
table
}
\description{
Parse a cefi opendap filename
}
