% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{parse_url}
\alias{parse_url}
\title{Parse a cefi URL into a table}
\usage{
parse_url(
  x = c("btm_htotal.nwa.full.hcast.daily.raw.r20230520.199301-201912.nc",
    "phycos.nwa.full.hcast.daily.regrid.r20230520.199301-201912.nc",
    "intppdiat.nwa.full.hcast.monthly.raw.r20230520.199301-201912.nc",
    "ffetot_btm.nwa.full.hcast.monthly.regrid.r20230520.199301-201912.nc")
)
}
\arguments{
\item{x}{one or more URLs}
}
\value{
table
}
\description{
Parse a cefi URL into a table
}
