% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_catalog}
\alias{read_catalog}
\title{Read the CEFI catalog}
\usage{
read_catalog(uri = catalog_uri(region = "NWA", xcast = "hindcast"))
}
\arguments{
\item{uri}{chr, the URI of the json resource}
}
\value{
table of metadata (of class CEFI_catalog)
}
\description{
Read the CEFI catalog
}
