% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{set_static}
\alias{set_static}
\title{Set the static attribute}
\usage{
set_static(x, y)
}
\arguments{
\item{x}{tidync object}

\item{y}{a static tidync object}
}
\value{
the input tidync object \code{x}
}
\description{
Set the static attribute
}
