% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{shift_stars}
\alias{shift_stars}
\title{Shift a stars object from (0,360) to (-180, 180) (or the other way)}
\usage{
shift_stars(x, to = c(180, 360)[1], name = "x")
}
\arguments{
\item{x}{a stars object}

\item{to}{num either 180 (meaning -180 to 180) or 360 (0 to 360)}

\item{name}{chr the name of the dimension along which the shift is made}
}
\description{
Shift a stars object from (0,360) to (-180, 180) (or the other way)
}
