% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static_lonlat}
\alias{static_lonlat}
\title{Get a static variables geolon and geolat}
\usage{
static_lonlat(x)
}
\arguments{
\item{x}{tidync object (with a static attribute object)}
}
\value{
a two element list of lon and lat
}
\description{
Get a static variables geolon and geolat
}
