% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static_open}
\alias{static_open}
\title{Open a static file}
\usage{
static_open(x = dplyr::slice(read_catalog(), 1))
}
\arguments{
\item{x}{a (single-row) table of "CEFI_catalog" class}
}
\value{
a tidync object
}
\description{
Open a static file
}
