% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static_url}
\alias{static_url}
\title{Retrieve static resource URL by region name}
\usage{
static_url(region = c("nwa", "nep")[1])
}
\arguments{
\item{region}{chr, one of "nwa" or "nep" (the latter is made up)}
}
\value{
character url for the opendap resource
}
\description{
Retrieve static resource URL by region name
}
