% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{tidync_as_stars}
\alias{tidync_as_stars}
\title{Cast a tidync object to stars}
\usage{
tidync_as_stars(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{other arguments (ignored)}
}
\value{
stars object
}
\description{
Taken from https://github.com/ropensci/tidync/issues/68#issuecomment-484773118
}
