% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{tidync_bbox}
\alias{tidync_bbox}
\title{Retrieve a bounding box ala \link[sf:st_bbox]{sf::st_bbox}}
\usage{
tidync_bbox(x)
}
\arguments{
\item{x}{tidync object}
}
\value{
\link[sf:st_bbox]{sf::st_bbox}
}
\description{
Retrieve a bounding box ala \link[sf:st_bbox]{sf::st_bbox}
}
