% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_uri}
\alias{catalog_uri}
\title{Retrieve a catalog based upon region and period}
\usage{
catalog_uri(
  period = c("history", "forecast")[1],
  region = c("NWA", "NEP")[1],
  stub = "https://psl.noaa.gov/cefi_portal/data_index"
)
}
\arguments{
\item{period}{chr one of "history" or "forecast"}

\item{region}{chr, one of "Northwest Atlantic", "NWA", "Northeast Pacific" or "NEP"}

\item{stub}{chr the base uri for catalogs}
}
\value{
character URL
}
\description{
Retrieve a catalog based upon region and period
}
