% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_stars}
\alias{cefi_stars}
\title{Extract data as stars}
\usage{
cefi_stars(
  x = cefi_open(),
  collapse_fun = mean,
  na.rm = TRUE,
  var = cefi_active(x)
)
}
\arguments{
\item{x}{tidync, likely filtered with hyper_filter}

\item{collapse_fun}{function reference for collapsing multiple layers or NULL to skip}

\item{na.rm}{logical, see \code{mean}}

\item{var}{chr, the variable to retrieve}
}
\value{
stars object
}
\description{
Extract data as stars
}
