% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_stars}
\alias{cefi_stars}
\title{Extract data as stars}
\usage{
cefi_stars(x = cefi_open())
}
\arguments{
\item{x}{tidync, likely filtered with hyper_filter}
}
\value{
stars object
}
\description{
Extract data as stars
}
