% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_subset}
\alias{cefi_subset}
\alias{[[.tidync_data}
\alias{[.tidync_data}
\title{Subset tidync_data objects}
\usage{
\method{[[}{tidync_data}(x, i)

\method{[}{tidync_data}(x, i)
}
\description{
Subset tidync_data objects
}
