% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_transforms}
\alias{cefi_transforms}
\title{Get the hyper transforms slightly doctored for time}
\usage{
cefi_transforms(x)
}
\arguments{
\item{x}{tidync or tidync_data object}

\item{list}{of one or more tidync axis transform(s)}
}
\description{
Get the hyper transforms slightly doctored for time
}
