% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cefi_var}
\alias{cefi_var}
\title{Get a CEFI variable as either a 'tidync_data' or 'stars' object}
\usage{
cefi_var(x = cefi_open(), form = c("tidync_data", "stars")[2])
}
\arguments{
\item{x}{the tidync object (possibly pre-filtered)}

\item{form}{one of 'tidync_data' or 'stars'}
}
\value{
either 'tidync_data' or 'stars' object
}
\description{
Get a CEFI variable as either a 'tidync_data' or 'stars' object
}
