% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{parse_catalog_uri}
\alias{parse_catalog_uri}
\title{Parse one or more URIs}
\usage{
parse_catalog_uri(
  uri = catalog_uri(),
  name_prefix = "cefi_data_indexing.Projects.CEFI.regional_mom6.cefi_portal."
)
}
\arguments{
\item{uri}{chr one or more catalog_uris}

\item{name_prefix}{chr the name prefix (of course)}
}
\value{
a table of parse table components
}
\description{
Parse one or more URIs
}
